# specs/007-monitor-web-ui/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Monitor UI - Alarm API
  description: Oracle DB 기반 알람 통계 및 최근 알람 목록 조회 API
  version: 1.0.0
  contact:
    name: JSScada Development Team

servers:
  - url: http://localhost:8000/api
    description: Local development server

tags:
  - name: Alarms
    description: 알람 조회 API

paths:
  /alarms/statistics:
    get:
      summary: 설비별 알람 통계 조회
      description: 17개 설비별 알람 합계와 일반 건수를 Oracle DB에서 조회합니다. 최근 24시간 동안 발생한 알람을 집계합니다.
      operationId: getAlarmStatistics
      tags:
        - Alarms
      responses:
        '200':
          description: 알람 통계 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment:
                    type: array
                    description: 설비별 알람 통계 배열 (17개)
                    minItems: 17
                    maxItems: 17
                    items:
                      $ref: '#/components/schemas/AlarmStatistics'
                  last_updated:
                    type: string
                    format: date-time
                    description: 마지막 갱신 시간
                    example: "2025-11-03T10:30:45.123Z"
              example:
                equipment:
                  - equipment_code: "KRCWO12ELOA101"
                    equipment_name: "Upper 로봇 용접"
                    alarm_count: 5
                    general_count: 3
                  - equipment_code: "KRCWO12ELOA102"
                    equipment_name: "Lower 로봇 용접"
                    alarm_count: 2
                    general_count: 1
                last_updated: "2025-11-03T10:30:45.123Z"
        '500':
          description: Oracle DB 연결 실패 또는 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Database connection failed"
                detail: "Unable to connect to Oracle DB"

  /alarms/recent:
    get:
      summary: 최근 알람 목록 조회
      description: 최근 발생한 알람을 시간 역순으로 조회합니다. 기본값은 5개이며 limit 파라미터로 조정 가능합니다.
      operationId: getRecentAlarms
      tags:
        - Alarms
      parameters:
        - name: limit
          in: query
          description: 조회할 알람 개수
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 100
            example: 5
      responses:
        '200':
          description: 최근 알람 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  alarms:
                    type: array
                    description: 알람 목록 (최신순)
                    items:
                      $ref: '#/components/schemas/Alarm'
              example:
                alarms:
                  - alarm_id: 12345
                    equipment_code: "KRCWO12ELOA101"
                    equipment_name: "Upper 로봇 용접"
                    alarm_type: "알람"
                    alarm_message: "온도 과열 경고"
                    occurred_at: "2025-11-03T10:28:15.000Z"
                  - alarm_id: 12344
                    equipment_code: "KRCWO12ELOA102"
                    equipment_name: "Lower 로봇 용접"
                    alarm_type: "일반"
                    alarm_message: "용접 전류 이상"
                    occurred_at: "2025-11-03T10:25:30.000Z"
        '400':
          description: 잘못된 요청 (limit 파라미터 범위 초과)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid parameter"
                detail: "limit must be between 1 and 100"
        '500':
          description: Oracle DB 연결 실패 또는 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Database connection failed"
                detail: "Unable to connect to Oracle DB"

components:
  schemas:
    AlarmStatistics:
      type: object
      description: 설비별 알람 통계
      required:
        - equipment_code
        - equipment_name
        - alarm_count
        - general_count
      properties:
        equipment_code:
          type: string
          description: 설비 코드 (14자리)
          pattern: '^[A-Z]{2}[A-Z]{3}\d{2}[A-Z]{3}[A-Z]\d{3}$'
          example: "KRCWO12ELOA101"
        equipment_name:
          type: string
          description: 설비 이름
          minLength: 1
          maxLength: 100
          example: "Upper 로봇 용접"
        alarm_count:
          type: integer
          description: 알람 합계 건수 (alarm_type='알람')
          minimum: 0
          example: 5
        general_count:
          type: integer
          description: 일반 건수 (alarm_type='일반')
          minimum: 0
          example: 3

    Alarm:
      type: object
      description: 알람 엔티티
      required:
        - alarm_id
        - equipment_code
        - equipment_name
        - alarm_type
        - alarm_message
        - occurred_at
      properties:
        alarm_id:
          type: integer
          description: 알람 고유 ID
          example: 12345
        equipment_code:
          type: string
          description: 설비 코드 (14자리)
          pattern: '^[A-Z]{2}[A-Z]{3}\d{2}[A-Z]{3}[A-Z]\d{3}$'
          example: "KRCWO12ELOA101"
        equipment_name:
          type: string
          description: 설비 이름
          minLength: 1
          maxLength: 100
          example: "Upper 로봇 용접"
        alarm_type:
          type: string
          description: 알람 유형
          enum: ["알람", "일반"]
          example: "알람"
        alarm_message:
          type: string
          description: 알람 메시지
          minLength: 1
          maxLength: 500
          example: "온도 과열 경고"
        occurred_at:
          type: string
          format: date-time
          description: 알람 발생 시간 (ISO 8601)
          example: "2025-11-03T10:28:15.000Z"

    Error:
      type: object
      description: 에러 응답
      required:
        - error
        - detail
      properties:
        error:
          type: string
          description: 에러 요약
          example: "Database connection failed"
        detail:
          type: string
          description: 에러 상세 내용
          example: "Unable to connect to Oracle DB"
        field:
          type: string
          description: 에러가 발생한 필드 (선택사항)
          example: "limit"